wip, workflow : submit, edit

+++ Sampoerna.EMS.Website/Models/LACK1/Lack1BaseItemModel.cs
@@ -2,7 +2,7 @@
 {
     public class Lack1BaseItemModel : BaseModel
     {
-        public long Lack1Id { get; set; }
+        public int Lack1Id { get; set; }
         public string Lack1Number { get; set; }
     }
 }
